Fix native binaries paths in docker environments
With the official ruby images from docker, svm_learn_native
and svm_classify_native were not found in gem specs.
This is normal because they are not specified as binaries
in the Gemspec.
This causes Gem.bin_path to fail.
The workaround is to find the path for the ruby binaries
svm_learn and svm_classify then append _native to the output.